~ updated entity attributes to use longs instead of ints to keep date/times in their values + added simple support for outcoming broadcasts to clients (idsOfInterest is not taken into account yet) ~ changed all the usages of DateTime.Now to DateTime.UtcNow + added support for sending notifications to the client + added support for scheduled tasks with TimerManager + added proper support for the skillQueue ~ fixed some oversights on the ItemAttribute class (some operators were not working properly, flags not being reset, etc) + added support for skill history + added support for SaveSkillQueue, CharStartTrainingSkillByTypeID, GetEndOfTraining and CharStopTrainingSkill to skillMgr bound service + added ItemGetInfo call to dogmaIM + added ItemChange keys for usage with item notifications

+++ Database/Server/04-entity.sql
@@ -33,7 +33,7 @@ DROP TABLE IF EXISTS `entity_attributes`;
 CREATE TABLE `entity_attributes` (
   `itemID` int(10) unsigned NOT NULL default '0',
   `attributeID` int(10) unsigned NOT NULL default '0',
-  `valueInt` int(10) unsigned default NULL,
+  `valueInt` bigint unsigned default NULL,
   `valueFloat` double unsigned default NULL,
   PRIMARY KEY  (`itemID`,`attributeID`),
   KEY `attributeID` (`attributeID`)
